Added in the concept of a sharer configuration
