+++ application/backend/.env.aws
@@ -14,4 +14,4 @@ ELSA_DATA_VERSION=local-dev-aws
 ELSA_DATA_BUILT=sometime
 ELSA_DATA_REVISION=0000000000000000000000000000000000000000
 
-ELSA_DATA_META_CONFIG_SOURCES="file('base') file('umccr-garvan-dev-super-admins') file('dev-localhost') file('datasets') aws-secret('ElsaDataLocalhost')"
+ELSA_DATA_META_CONFIG_SOURCES="file('base') file('umccr-garvan-dev-super-admins') file('dev-localhost') file('datasets') file('sharers') aws-secret('ElsaDataLocalhost')"
