+++ application/backend/.env.linux
@@ -14,4 +14,4 @@ ELSA_DATA_VERSION=local-dev-linux
 ELSA_DATA_BUILT=sometime
 ELSA_DATA_REVISION=0000000000000000000000000000000000000000
 
-ELSA_DATA_META_CONFIG_SOURCES="file('base') file('umccr-garvan-dev-super-admins') file('dev-localhost') file('datasets') linux-pass('elsa-data')"
+ELSA_DATA_META_CONFIG_SOURCES="file('base') file('umccr-garvan-dev-super-admins') file('dev-localhost') file('datasets') file('sharers') linux-pass('elsa-data')"
