+++ application/backend/.env.mac
@@ -14,4 +14,4 @@ ELSA_DATA_VERSION=local-dev-mac
 ELSA_DATA_BUILT=sometime
 ELSA_DATA_REVISION=0000000000000000000000000000000000000000
 
-ELSA_DATA_META_CONFIG_SOURCES="file('base') file('umccr-garvan-dev-super-admins') file('dev-localhost') file('datasets') osx-keychain('elsa-data')"
+ELSA_DATA_META_CONFIG_SOURCES="file('base') file('umccr-garvan-dev-super-admins') file('dev-localhost') file('datasets') file('sharers') osx-keychain('elsa-data')"
