+++ application/backend/config/base.json5
@@ -3,7 +3,4 @@
   ontoFhirUrl: "https://onto.prod.umccr.org/fhir",
   maxmindDbAssetPath: "asset/maxmind/db/",
   releaseKeyPrefix: "R",
-  htsget: {
-    url: "https://htsget.elsa.dev.umccr.org",
-  },
 }
