+++ application/backend/config/dev-deployed.json5
@@ -4,7 +4,4 @@
   logger: {
     level: "debug",
   },
-  htsget: {
-    url: "https://htsget.elsa.dev.umccr.org",
-  },
 }
