+++ application/backend/src/app-router.ts
@@ -8,6 +8,7 @@ import { datasetRouter } from "./api/routes/internal-trpc/dataset-router";
 import { releaseRouter } from "./api/routes/internal-trpc/release-router";
 import { dacRouter } from "./api/routes/internal-trpc/dac-router";
 import { auditEventRouter } from "./api/routes/internal-trpc/audit-event-router";
+import { sharerRouter } from "./api/routes/internal-trpc/sharer-router";
 
 export const appRouter = router({
   dac: dacRouter,
@@ -17,6 +18,7 @@ export const appRouter = router({
   releaseJob: releaseJobRouter,
   releaseRouter: releaseRouter,
   releaseDataEgress: releaseDataEgressRouter,
+  sharer: sharerRouter,
   test: testRouter,
   user: userRouter,
 });
