+++ application/backend/src/bootstrap-settings.ts
@@ -1,14 +1,44 @@
-import { BaseClient, Issuer } from "openid-client";
-import { writeFile } from "fs/promises";
-import * as temp from "temp";
+import { Issuer } from "openid-client";
 import { ElsaSettings } from "./config/elsa-settings";
 import _ from "lodash";
-import { ElsaConfiguration } from "./config/config-constants";
-import { locateHtmlDirectory } from "./app-helpers";
+import { ElsaConfigurationType } from "./config/config-schema";
 import * as path from "path";
+import { OidcType } from "./config/config-schema-oidc";
+import { SharerType } from "./config/config-schema-sharer";
+
+/**
+ * Enrich the raw OIDC configuration by creating an Issuer
+ * instance.
+ *
+ * @param allowEmptyIssuer
+ * @param oidcConfiguration
+ */
+export async function oidcConfigurationToSettings(
+  allowEmptyIssuer: boolean,
+  oidcConfiguration?: OidcType
+) {
+  if (!oidcConfiguration) return undefined;
+
+  let issuer: Issuer | undefined = undefined;
+
+  if (oidcConfiguration.issuerUrl) {
+    issuer = await Issuer.discover(oidcConfiguration.issuerUrl);
+  } else {
+    if (!allowEmptyIssuer)
+      throw new Error(
+        "Only localhost development launches can exist without setting up an OIDC issuer"
+      );
+  }
+
+  return {
+    issuer: issuer,
+    clientId: oidcConfiguration.clientId,
+    clientSecret: oidcConfiguration.clientSecret,
+  };
+}
 
 export async function bootstrapSettings(
-  config: ElsaConfiguration
+  config: ElsaConfigurationType
 ): Promise<ElsaSettings> {
   // we now have our 'config' - which is the plain text values from all our configuration sources..
   // however our 'settings' are more than that - the settings involve things that need to be
@@ -16,7 +46,7 @@ export async function bootstrapSettings(
   // Settings might involve writing a cert file to disk and setting a corresponding env variable.
   // Settings might involve doing OIDC discovery and then using the returned value
 
-  const rootCa: string | undefined = config.edgeDb?.tlsRootCa;
+  /*const rootCa: string | undefined = config.edgeDb?.tlsRootCa;
 
   if (rootCa) {
     // edge db requires a TLS connection - and so we need to construct our own
@@ -26,7 +56,7 @@ export async function bootstrapSettings(
 
     // We do not yet have a logger available at this point in the bootstrapper
     // So only enable this console logging if debugging the CA config
-    // (in general this mechanism has been working well since initial implementation)
+    // (in general this mechanism has been notWorking well since initial implementation)
     //console.log(
     //  `Discovered TLS Root CA configuration so constructing CA on disk at ${rootCaLocation} and setting path in environment variable EDGEDB_TLS_CA_FILE`
     //);
@@ -35,7 +65,7 @@ export async function bootstrapSettings(
     await writeFile(rootCaLocation, rootCa.replaceAll("\\n", "\n"));
 
     process.env["EDGEDB_TLS_CA_FILE"] = rootCaLocation;
-  }
+  } */
 
   const logoPath = _.get(config, "branding.logoPath");
   let logoUriRelative: string | undefined;
@@ -65,17 +95,6 @@ export async function bootstrapSettings(
       );
   }
 
-  let issuer: Issuer | undefined = undefined;
-
-  if (_.has(config, "oidc")) {
-    issuer = await Issuer.discover(config.oidc?.issuerUrl!);
-  } else {
-    if (!isDevelopment || !isLocalhost)
-      throw new Error(
-        "Only localhost development launches can exist without setting up an OIDC issuer"
-      );
-  }
-
   let loggerTransportTargets: any[] = _.get(config, "logger.transportTargets");
 
   // grab the targets from our config - but default to a sensible default that just logs to stdout
@@ -91,17 +110,26 @@ export async function bootstrapSettings(
 
   if (logLevel) loggerTransportTargets.forEach((l) => (l.level ??= logLevel));
 
-  // TODO fix this logic once we have a few more AWS settings and know what is required
-  const hasAws =
-    config.aws?.signingAccessKeyId ||
-    config.aws?.signingSecretAccessKey ||
-    config.aws?.tempBucket;
-  const hasAwsSigning =
-    _.get(config, "aws.signingAccessKeyId") &&
-    _.get(config, "aws.signingSecretAccessKey");
+  const hasAws = config.aws?.tempBucket;
+
   const hasCloudflare =
     _.get(config, "cloudflare.signingAccessKeyId") &&
     _.get(config, "cloudflare.signingAccessKeyId");
+
+  // whilst it is probably needed eventually that we support multiple sharers of each type
+  // (for instance to support multiple htsget endpoints) - for the moment we have not thought
+  // this through and hence limit it to one of each
+  const sharers: SharerType[] = _.get(config, "sharers");
+  const sharerTypes = sharers.map((s) => s.type);
+
+  const sharerDuplicates = sharerTypes.filter(
+    (item, index) => sharerTypes.indexOf(item) !== index
+  );
+
+  if (sharerDuplicates.length > 0)
+    throw new Error(
+      `For the moment, only a single sharer of each type can be specified. The following sharers types are duplicated -> ${sharerDuplicates}`
+    );
 
   return {
     deployedUrl: deployedUrl,
@@ -109,32 +137,13 @@ export async function bootstrapSettings(
       _.get(config, "serviceDiscoveryNamespace") ?? "elsa-data",
     host: _.get(config, "host"),
     port: _.get(config, "port"),
-    mailer: {
-      mode: _.get(config, "mailer.mode"),
-      maxConnections: _.get(config, "mailer.maxConnections"),
-      sendingRate: _.get(config, "mailer.sendingRate"),
-      options: _.get(config, "mailer.options"),
-      defaults: _.get(config, "mailer.defaults"),
-    },
-    oidcClientId: _.get(config, "oidc.clientId")!,
-    oidcClientSecret: _.get(config, "oidc.clientSecret")!,
-    // NOTE this can be undefined for dev localhost - in which case any OIDC code paths will fail
-    // but we expect that
-    oidcIssuer: issuer!,
-    htsget: _.get(config, "htsget.url")
-      ? {
-          maxAge: _.get(config, "htsget.maxAge"),
-          url: new URL(_.get(config, "htsget.url")),
-        }
-      : undefined,
+    mailer: config.mailer ? _.get(config, "mailer") : undefined,
+    oidc: await oidcConfigurationToSettings(
+      isDevelopment || isLocalhost,
+      config.oidc
+    ),
     aws: hasAws
       ? {
-          signingAccessKeyId: hasAwsSigning
-            ? _.get(config, "aws.signingAccessKeyId")
-            : undefined,
-          signingSecretAccessKey: hasAwsSigning
-            ? _.get(config, "aws.signingSecretAccessKey")
-            : undefined,
           tempBucket: _.get(config, "aws.tempBucket"),
         }
       : undefined,
@@ -150,6 +159,7 @@ export async function bootstrapSettings(
     sessionSecret: _.get(config, "session.secret")!,
     sessionSalt: _.get(config, "session.salt")!,
     dacs: _.get(config, "dacs"),
+    sharers: sharers,
     logger: {
       name: "elsa-data",
       level: logLevel,
@@ -197,8 +207,7 @@ export async function bootstrapSettings(
       logoPath: logoPath,
       logoUriRelative: logoUriRelative,
     },
-    // these are our special arrays that can be constructed with + and - definitions
-    datasets: (config.datasets as any[]) ?? [],
-    superAdmins: (config.superAdmins as any[]) ?? [],
+    datasets: _.get(config, "datasets"),
+    superAdmins: _.get(config, "superAdmins"),
   };
 }
