+++ application/backend/src/config/config-schema.ts
@@ -1,300 +1,185 @@
-import { parseMeta } from "./meta/meta-parser";
-import { ProviderAwsSecretsManager } from "./providers/provider-aws-secrets-manager";
-import { ProviderGcpSecretsManager } from "./providers/provider-gcp-secrets-manager";
-import { ProviderFile } from "./providers/provider-file";
-import { configZodDefinition, ElsaConfiguration } from "./config-constants";
-import { ProviderOsxKeychain } from "./providers/provider-osx-keychain";
-import { ProviderLinuxPass } from "./providers/provider-linux-pass";
-import jsonpath from "jsonpath";
-import _ from "lodash";
+import { z } from "zod";
+import { Sensitive } from "./config-schema-sensitive";
+import { DacSchema } from "./config-schema-dac";
+import { DatasetSchema } from "./config-schema-dataset";
+import { SharerSchema } from "./config-schema-sharer";
+import { MailerSchema } from "./config-schema-mailer";
+import { BrandingSchema } from "./config-schema-branding";
+import { OidcSchema } from "./config-schema-oidc";
 
-const env_prefix = "ELSA_DATA_CONFIG_";
+export const CONFIG_SOURCES_ENVIRONMENT_VAR = `ELSA_DATA_META_CONFIG_SOURCES`;
+export const CONFIG_FOLDERS_ENVIRONMENT_VAR = `ELSA_DATA_META_CONFIG_FOLDERS`;
 
 /**
- * Process/merge in a new configuration file - but with some special helper logic that allows array
- * manipulation and JSONpath queries.
- *
- * @param startingConfig the starting configuration that we are looking to apply array options to
- * @param newConfig a new configuration that may or may not have special logic fields (at the top level)
- * @param arrayMarkers an array of "key names" that we want to enable this array functionality for (e.g. ["datasets", "admins"])
+ * The master schema definition for our configuration objects (i.e. JSON).
  */
-function mergeNewConfig(
-  startingConfig: any,
-  newConfig: any,
-  arrayMarkers: string[]
-): any {
-  // start with the intention of returning back exactly what we started with
-  // we will now mutate this
-  const returnConfig = _.cloneDeep(startingConfig);
-
-  // we don't want to accidentally refer to this so null it out
-  startingConfig = null;
-
-  // we want to allow 'deletion' of config entries using a pretty loose logic where they can specify anything
-  // that looks like an id and we broadly search for it
-  const idPresent = (possibleId: string, obj: any): boolean => {
-    for (const [k, v] of Object.entries(obj)) {
-      if (k === "uri" && v === possibleId) return true;
-      if (k === "id" && v === possibleId) return true;
-      if (k === "name" && v === possibleId) return true;
-      if (k === "sub" && v === possibleId) return true;
-    }
-
-    return false;
-  };
-
-  // our first step is we want to apply any of the "array modifier" keys we find in the new config
-  // these only are allowed as top-level keys
-
-  for (const am of arrayMarkers) {
-    const addKey = `+${am}`;
-    const deleteKey = `-${am}`;
-    const replaceKey = am;
-
-    if (replaceKey in newConfig) {
-      // the presence of the key name directly means we are doing a straight "replace" - and hence can't be doing an add or delete
-      if (addKey in newConfig || deleteKey in newConfig) {
-        throw new Error(
-          `If a configuration has a key '${replaceKey}' to replace array content - it does not make any sense to also have a '${addKey}' or '${deleteKey}' key`
-        );
-      }
-
-      returnConfig[am] = newConfig[am];
-
-      delete newConfig[replaceKey];
-    } else {
-      // it may make sense to have BOTH a + and a - so we need to be cool with both appearing
-
-      if (addKey in newConfig) {
-        returnConfig[am].push(...newConfig[addKey]);
-
-        // strip out the +?? key as we have handled it and we don't want is handled by lodash merge
-        delete newConfig[addKey];
-      }
-
-      if (deleteKey in newConfig) {
-        for (const toDelete of newConfig[deleteKey]) {
-          const startLength = returnConfig[am].length;
-          returnConfig[am] = returnConfig[am].filter(
-            (v: any) => !idPresent(toDelete, v)
-          );
-          if (startLength === returnConfig[am].length) {
-            throw new Error(
-              `The configuration instruction ${deleteKey} of ${toDelete} did not do anything as no existing config had an entry with an id,uri or name of ${toDelete}`
-            );
-          }
-        }
-
-        // strip out the -?? key as we have handled it and we don't want is handled by lodash merge
-        delete newConfig[deleteKey];
-      }
-    }
-  }
-
-  // our top-level keys can be complex JSON path expressions that resolve to a single value
-  // this lets us set password fields in arrays for instance
-  for (const [key, value] of Object.entries(newConfig)) {
-    // if the key is just regular alphanumeric then we can let it be handled normally
-    if (/^[A-Za-z0-9]*$/.test(key)) continue;
-
-    // find our what already exists according to this complex query
-    const matching = jsonpath.query(returnConfig, key);
-
-    // we do not want to use this mechanism for replacing multiple nodes
-    if (matching.length > 1)
-      throw new Error(
-        `Configuration handling for complex key ->${key}<- resulted in a path value that was not a single string or number - which is incompatible with how this mechanism is designed to be used`
-      );
-
-    if (matching.length === 0) {
-      // if the value is not already present - jsonpath apply can't really help us out because it only applies to existing objects
-      throw new Error(
-        `Configuration handling for complex key ->${key}<- can _currently_ only replace existing fields, not create them new - so you may need to add an empty field in an earlier configuration i.e a blank secret ""`
-      );
-    } else {
-      // if there are already values present matching this key - they cannot be objects or arrays
-      if (
-        !_.isString(matching[0]) &&
-        !_.isNumber(matching[0]) &&
-        !_.isNil(matching[0])
+export const configZodDefinition = z.object({
+  serviceDiscoveryNamespace: z
+    .optional(z.string())
+    .default("elsa-data")
+    .describe(
+      "The name of the root artifact used for dynamic service discovery - this will differ for each deployment environment. For AWS - it refers to a CloudMap namespace"
+    ),
+  oidc: z.optional(OidcSchema),
+  session: z.optional(
+    z.object({
+      secret: z
+        .string()
+        .describe(
+          "A long string secret that is used as part of the session encryption"
+        )
+        .brand<Sensitive>(),
+      salt: z
+        .string()
+        .length(16)
+        .describe(
+          "A 16 character salt that is used as part of the session encryption"
+        )
+        .brand<Sensitive>(),
+    })
+  ),
+  aws: z.optional(
+    z.object({
+      tempBucket: z
+        .optional(z.string())
+        .describe(
+          "A bucket that can be used for storing temporary artifacts - can have a Lifecycle that removes files after a day"
+        ),
+    })
+  ),
+  cloudflare: z.optional(
+    z.object({
+      signingAccessKeyId: z
+        .string()
+        .describe(
+          "A CloudFlare R2 access key id for a user with read permission of files that can be shared via signed URLs"
+        )
+        .brand<Sensitive>(),
+      signingSecretAccessKey: z
+        .string()
+        .describe(
+          "A CloudFlare R2 secret access key for a user with read permission of files that can be shared via signed URLs"
+        )
+        .brand<Sensitive>(),
+    })
+  ),
+  // the rate limiting options are basically a pass through to the Fastify rate limit plugin
+  // for the moment we have picked only a subset of the full configuration items
+  rateLimit: z.optional(
+    z.object({
+      allowList: z
+        .optional(z.array(z.string()))
+        .describe(
+          "The IP addresses of any clients allowed to bypass rate limiting"
+        ),
+      max: z
+        .optional(z.number().int())
+        .describe(
+          "The maximum number of API calls allowed from a single IP in the time window"
+        ),
+      timeWindow: z
+        .optional(z.number().int())
+        .describe(
+          "The number of milliseconds in the time window in which rate limit events are measured"
+        ),
+    })
+  ),
+  logger: z
+    .object({
+      level: z
+        .string()
+        .describe(
+          "The logging level as per Pino (all the standard level strings + silent)"
+        )
+        .default("debug"),
+      transportTargets: z
+        .array(
+          z
+            .object({
+              target: z.string(),
+            })
+            .passthrough()
+        )
+        .describe("An array of Pino logger transport targets configurations")
+        .default([]),
+    })
+    .default({
+      level: "debug",
+      transportTargets: [],
+    }),
+  ontoFhirUrl: z.optional(z.string()),
+  // all production deployments will require this to be set - though the logic for this check is elsewhere
+  deployedUrl: z.optional(
+    z
+      .string()
+      .describe(
+        "The externally accessible Url for the deployed location of Elsa Data"
       )
-        throw new Error(
-          `Configuration handling for complex key ->${key}<- resulted in a path value that was not a single string or number - which is incompatible with how this mechanism is designed to be used`
-        );
+  ),
+  maxmindDbAssetPath: z
+    .optional(z.string())
+    .describe("The path to the Maxmind city database"),
+  releaseKeyPrefix: z
+    .string()
+    .default("R")
+    .describe("The prefix appended on the releaseKey before the count number"),
+  host: z
+    .string()
+    .ip()
+    .default("0.0.0.0")
+    .describe("The host interface to bind to"),
+  port: z.coerce
+    .number()
+    .positive()
+    .int()
+    .default(3000)
+    .describe("The port to bind to"),
+  datasets: z
+    .array(DatasetSchema)
+    .default([])
+    .describe(
+      "An array defining the datasets which are shareable from this instance"
+    ),
+  superAdmins: z
+    .array(
+      z.object({
+        sub: z
+          .string()
+          .describe(
+            "The subject id of the user that should be given super admin permissions"
+          ),
+      })
+    )
+    .default([])
+    .describe(
+      "A collection of users with super administration rights i.e. the ability to alter other user rights"
+    ),
+  dacs: z.array(DacSchema).default([
+    {
+      id: "manual",
+      type: "manual",
+      description: "Manual",
+    },
+  ]),
+  sharers: z
+    .array(SharerSchema)
+    .default([])
+    .describe(
+      "An array defining the sharing mechanisms which are to be enabled from this instance"
+    ),
+  // if present, a mailer is being configured and if not present, then the mailer does not start
+  mailer: z.optional(MailerSchema),
+  devTesting: z.optional(
+    z.object({
+      allowTestUsers: z
+        .boolean()
+        .describe(
+          "If test users should be allowed, including various techniques used to adjust user sessions"
+        ),
+      allowTestRoutes: z.boolean().describe("If test routes should be added"),
+    })
+  ),
+  branding: z.optional(BrandingSchema),
+});
 
-      // the value we want to set cannot be anything but a string/number (we could probably check this earlier)
-      if (!_.isString(value) && !_.isNumber(value))
-        throw new Error(
-          `Configuration handling for complex key ->${key}<- resulted in a value to be set that was not a single string or number - which is incompatible with how this mechanism is designed to be used`
-        );
-
-      // now that we have ensured they aren't getting up to anything funny - we apply the replacement of the single value
-      jsonpath.apply(returnConfig, key, (x) => value);
-    }
-
-    delete newConfig[key];
-  }
-
-  // all the "special" fields have been removed from newConfig and we have applied their logic ourselves
-  // now we can just use the basis Lodash merge functionality for the rest of the newConfig
-  return _.merge(returnConfig, newConfig);
-}
-
-/**
- * Give the raw JSON of configuration, return the data loaded into Convict.
- * This function is drop in replacement for getMetaConfig - but is useful in
- * some testing situations.
- *
- * @param config
- */
-export async function getDirectConfig(config: any): Promise<ElsaConfiguration> {
-  return configZodDefinition.strict().parse(config);
-}
-
-/**
- * Given a meta configuration description (a string listing a sequence of config
- * providers), this loads all the relevant configs in order and returns them
- * merged and loaded via Zod schema checking. Env variables will be late bound
- * to be able to also set config values.
- *
- * @param meta a string of source of configuration information e.g "file(a) aws-secret(b)"
- */
-export async function getMetaConfig(meta: string): Promise<ElsaConfiguration> {
-  // the meta syntax tells us where to source configuration from and in what order
-  const metaProviders = parseMeta(meta);
-
-  // the raw config JSON as loaded from the sources
-  const rawConfigs = [];
-
-  for (const mp of metaProviders) {
-    let providerConfig;
-    switch (mp.providerToken.value) {
-      case "aws-secret":
-        providerConfig = await new ProviderAwsSecretsManager(
-          mp.argTokens
-        ).getConfig();
-        break;
-      case "gcloud-secret":
-        providerConfig = await new ProviderGcpSecretsManager(
-          mp.argTokens
-        ).getConfig();
-        break;
-      case "file":
-        providerConfig = await new ProviderFile(mp.argTokens).getConfig();
-        break;
-      case "osx-keychain":
-        providerConfig = await new ProviderOsxKeychain(
-          mp.argTokens
-        ).getConfig();
-        break;
-      case "linux-pass":
-        providerConfig = await new ProviderLinuxPass(mp.argTokens).getConfig();
-        break;
-      default:
-        throw new Error(
-          `unrecognised configuration provider ${mp.providerToken.value}`
-        );
-    }
-    rawConfigs.push(providerConfig);
-  }
-
-  let configObject: ElsaConfiguration = {} as ElsaConfiguration;
-
-  // one by one merge each config in order into the config object
-  for (const initialRawConfig of rawConfigs) {
-    configObject = mergeNewConfig(configObject, initialRawConfig, [
-      "superAdmins",
-      "datasets",
-      "dacs",
-    ]);
-  }
-
-  // generalise this - but for proper functioning of other aspects of the DAC service - we must ensure
-  // that these are unique within the instance
-  {
-    const idSet = new Set<string>();
-    for (const c of configObject.dacs || []) {
-      idSet.add(c.id);
-    }
-    if (idSet.size !== (configObject.dacs || []).length)
-      throw new Error(
-        "The 'id' field of the configuration 'dacs' array must be unique"
-      );
-  }
-
-  // similarly things will go horribly wrong if dataset uris are repeated
-  {
-    const uriSet = new Set<string>();
-    for (const c of configObject.datasets || []) {
-      uriSet.add(c.uri);
-    }
-    if (uriSet.size !== (configObject.datasets || []).length)
-      throw new Error(
-        "The 'uri' field of the configuration 'datasets' array must be unique"
-      );
-  }
-
-  const trySetEnvironmentVariableString = (
-    env_suffix: string,
-    path: string
-  ) => {
-    const v = process.env[`${env_prefix}${env_suffix}`];
-    if (v) _.set(configObject, path, v);
-  };
-  const trySetEnvironmentVariableInteger = (
-    env_suffix: string,
-    path: string
-  ) => {
-    const v = process.env[`${env_prefix}${env_suffix}`];
-    if (v) _.set(configObject, path, parseInt(v));
-  };
-
-  // TODO a mechanism that perhaps _parses_ env names - and works out the path itself..
-  // (i.e. replace _ with . and capitalise! but better)
-
-  trySetEnvironmentVariableString(
-    "SERVICE_DISCOVERY_NAMESPACE",
-    "serviceDiscoveryNamespace"
-  );
-  trySetEnvironmentVariableString("SESSION_SECRET", "session.secret");
-  trySetEnvironmentVariableString("SESSION_SALT", "session.salt");
-  trySetEnvironmentVariableString(
-    "AWS_SIGNING_ACCESS_KEY_ID",
-    "aws.signingAccessKeyId"
-  );
-  trySetEnvironmentVariableString(
-    "AWS_SIGNING_SECRET_ACCESS_KEY",
-    "aws.signingSecretAccessKey"
-  );
-  trySetEnvironmentVariableString("AWS_TEMP_BUCKET", "aws.tempBucket");
-  trySetEnvironmentVariableString(
-    "CLOUDFLARE_SIGNING_ACCESS_KEY_ID",
-    "cloudflare.signingAccessKeyId"
-  );
-  trySetEnvironmentVariableString(
-    "CLOUDFLARE_SIGNING_SECRET_ACCESS_KEY",
-    "cloudflare.signingSecretAccessKey"
-  );
-  trySetEnvironmentVariableString("DEPLOYED_URL", "deployedUrl");
-
-  trySetEnvironmentVariableInteger("PORT", "port");
-  trySetEnvironmentVariableString("HOST", "host");
-
-  trySetEnvironmentVariableString("LOGGER_LEVEL", "logger.level");
-
-  trySetEnvironmentVariableString("MAILER_OPTIONS", "mailer.options");
-  trySetEnvironmentVariableString("MAILER_DEFAULTS", "mailer.defaults");
-  trySetEnvironmentVariableString("MAILER_MODE", "mailer.mode");
-
-  trySetEnvironmentVariableInteger("HTSGET_MANIFEST_TTL", "htsget.manifestTtl");
-  trySetEnvironmentVariableString("HTSGET_URL", "htsget.url");
-
-  // perform validation on the final config object and return it transformed
-  // (the transform will do type coercion and setting defaults)
-  try {
-    return configZodDefinition.strict().parse(configObject);
-  } catch (e) {
-    console.warn(`failed strict parsing: ${e}`);
-    return configZodDefinition.parse(configObject);
-  }
-}
+export type ElsaConfigurationType = z.infer<typeof configZodDefinition>;
