+++ application/backend/src/config/elsa-settings.ts
@@ -2,8 +2,13 @@ import { Issuer } from "openid-client";
 import { RateLimitPluginOptions } from "@fastify/rate-limit";
 import SMTPTransport from "nodemailer/lib/smtp-transport";
 import { LoggerOptions } from "pino";
-import { Dac } from "./config-schema-dac";
+import { DacType } from "./config-schema-dac";
 import { DatasetType } from "./config-schema-dataset";
+import { SharerType } from "./config-schema-sharer";
+import { ElsaConfigurationType } from "./config-schema";
+import { MailerType } from "./config-schema-mailer";
+import { BrandingType } from "./config-schema-branding";
+import { OidcType } from "./config-schema-oidc";
 
 /**
  * The rich, well-typed settings for Elsa.
@@ -15,13 +20,6 @@ export type ElsaSettings = {
 
   host: string;
   port: number;
-  mailer: {
-    mode: "None" | "SES" | "SMTP";
-    maxConnections?: number | undefined;
-    sendingRate?: number | undefined;
-    options?: SMTPTransport.Options | string;
-    defaults?: any;
-  };
 
   // the namespace in which we should be doing service discovery for dynamic services
   serviceDiscoveryNamespace: string;
@@ -29,26 +27,11 @@ export type ElsaSettings = {
   sessionSecret: string;
   sessionSalt: string;
 
-  oidcIssuer: Issuer;
-  oidcClientId: string;
-  oidcClientSecret: string;
+  oidc?: Omit<OidcType, "issuerUrl"> & { issuer?: Issuer };
 
-  dacs: Dac[];
-
-  logger: LoggerOptions;
-
-  htsget?: {
-    maxAge: number;
-    url: URL;
-  };
-
-  // optional details to allow sharing of objects in AWS
+  // details that are required if running in AWS
   aws?: {
-    // if using AWS then temp bucket is required
     tempBucket: string;
-    // it is possible to use AWS but not necessarily use signing
-    signingAccessKeyId?: string;
-    signingSecretAccessKey?: string;
   };
 
   // optional signing details to allow sharing of objects in CloudFlare R2
@@ -80,11 +63,6 @@ export type ElsaSettings = {
     sub: string;
   }[];
 
-  datasets: DatasetType[];
-
-  // options to pass into the rate limiter
-  rateLimit: RateLimitPluginOptions;
-
   // dev/testing settings that can be specified as long as the NODE_ENV is development
   // if NODE_ENV is production then this the presence of any configuration leading to this
   // will fail to launch
@@ -97,9 +75,21 @@ export type ElsaSettings = {
     allowTestRoutes: boolean;
   };
 
-  branding?: {
-    brandName?: string;
-    logoPath?: string;
-    logoUriRelative?: string;
-  };
+  // pass through directly from configuration - eventually we want to pass everything through
+  // directly and essentially remove ElsaSettings as a type
+
+  dacs: DacType[];
+
+  logger: LoggerOptions;
+
+  datasets: DatasetType[];
+
+  sharers: SharerType[];
+
+  // options to pass into the rate limiter
+  rateLimit: RateLimitPluginOptions;
+
+  mailer?: MailerType;
+
+  branding?: BrandingType & { logoUriRelative?: string };
 };
