+++ application/backend/src/config/providers/provider-file.ts
@@ -3,7 +3,7 @@ import { constants, promises as fs } from "fs";
 import path, { resolve } from "path";
 import { ProviderBase } from "./provider-base";
 import json5 from "json5";
-import { CONFIG_FOLDERS_ENVIRONMENT_VAR } from "../config-constants";
+import { CONFIG_FOLDERS_ENVIRONMENT_VAR } from "../config-schema";
 
 /**
  * A provider that can get config from a file
