+++ application/frontend/src/providers/logged-in-user-config-relay-provider.tsx
@@ -1,15 +1,19 @@
 import React from "react";
 import { createCtx } from "./create-ctx";
-import { Dac } from "../../../backend/src/config/config-schema-dac";
+import { DacType } from "../../../backend/src/config/config-schema-dac";
 import { trpc } from "../helpers/trpc";
 import { useLoggedInUser } from "./logged-in-user-provider";
+import { SharerType } from "../../../backend/src/config/config-schema-sharer";
 
 export type LoggedInUserConfigRelay = {
   // the set of datasets currently available from the instance
   datasets: Record<string, string>;
 
+  // the set of sharers currently available from the instance
+  sharers: SharerType[];
+
   // the set of DACS currently available from the instance (or [] if this user cannot create releases)
-  dacs: Dac[];
+  dacs: DacType[];
 };
 
 /**
@@ -26,24 +30,28 @@ export const LoggedInUserConfigRelayProvider: React.FC<Props> = (
 ) => {
   const loggedInUser = useLoggedInUser();
 
-  const datasetsQuery = trpc.datasetRouter.getConfiguredDatasets.useQuery(
-    undefined,
-    {
-      refetchOnWindowFocus: false,
-      refetchOnReconnect: false,
-      retry: false,
-      enabled: !!loggedInUser,
-    }
-  );
-  const dacQuery = trpc.dac.getConfiguredDacs.useQuery(undefined, {
+  const qSettings = {
     refetchOnWindowFocus: false,
     refetchOnReconnect: false,
     retry: false,
     enabled: !!loggedInUser,
-  });
+  };
+
+  const datasetsQuery = trpc.datasetRouter.getConfiguredDatasets.useQuery(
+    undefined,
+    qSettings
+  );
+
+  const sharersQuery = trpc.sharer.getConfiguredSharers.useQuery(
+    undefined,
+    qSettings
+  );
+
+  const dacQuery = trpc.dac.getConfiguredDacs.useQuery(undefined, qSettings);
 
   const val = {
     datasets: datasetsQuery.isSuccess ? datasetsQuery.data : {},
+    sharers: sharersQuery.isSuccess ? sharersQuery.data : [],
     dacs: dacQuery.isSuccess ? dacQuery.data : [],
   };
